Java 5 type checking and compiler warnings
git-svn-id: svn://source.pentaho.org/svnkettleroot/Kettle/trunk@3719 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src/org/pentaho/di/core/widget/TextVar.java
@@ -170,7 +170,7 @@ public class TextVar extends Composite
             public void keyPressed(KeyEvent e)
             {
                 // CTRL-<SPACE> --> Insert a variable
-                if ((int)e.character == ' ' && (( e.stateMask&SWT.CONTROL)!=0) && (( e.stateMask&SWT.ALT)==0) ) 
+                if (e.character == ' ' && (( e.stateMask&SWT.CONTROL)!=0) && (( e.stateMask&SWT.ALT)==0) ) 
                 { 
                     Event event = new Event();
                     event.widget = textField;
